refactor: extract user management layout to layouts/user_management.py
